+++ packages/illustrations/src/__generated__/pictogram/types/PictogramName.ts
@@ -24,6 +24,7 @@ export type PictogramName =
   | 'apartOfDropsNft'
   | 'applyForHigherLimits'
   | 'apyInterest'
+  | 'arrowsUpDown'
   | 'assetEncryption'
   | 'assetHubNavigation'
   | 'assetManagement'