+++ packages/mobile/src/dates/__stories__/DatePicker.stories.tsx
@@ -97,10 +97,10 @@ export const FullExample = () => {
           accessibilityLabel="Birthdate"
           calendarIconButtonAccessibilityLabel="Birthdate calendar"
           labelNode={
-            <HStack alignItems="center" gap={1}>
+            <HStack alignItems="center">
               <InputLabel>Birthdate</InputLabel>
               <Tooltip content="This will be visible to other users.">
-                <Icon color="fgMuted" name="info" size="xs" />
+                <Icon active color="fg" name="info" padding={0.75} size="xs" />
               </Tooltip>
             </HStack>
           }
@@ -139,10 +139,10 @@ export const CustomLabel = () => {
             accessibilityLabel="Date of birth"
             calendarIconButtonAccessibilityLabel="Date of birth calendar"
             labelNode={
-              <HStack alignItems="center" gap={1}>
+              <HStack alignItems="center">
                 <InputLabel>Date of birth</InputLabel>
                 <Tooltip content="This will be visible to other users.">
-                  <Icon color="fgMuted" name="info" size="xs" />
+                  <Icon active color="fg" name="info" padding={0.75} size="xs" />
                 </Tooltip>
               </HStack>
             }