+++ packages/web/src/dates/__stories__/DatePicker.stories.tsx
@@ -73,14 +73,12 @@ export const Examples = () => {
           {...props}
           accessibilityLabel="Date of birth"
           labelNode={
-            <InputLabel>
-              <HStack alignItems="center" gap={1}>
-                Date of birth
-                <Tooltip content="This will be visible to other users.">
-                  <Icon active color="fg" name="info" size="xs" tabIndex={0} />
-                </Tooltip>
-              </HStack>
-            </InputLabel>
+            <HStack alignItems="center">
+              <InputLabel>Date of birth</InputLabel>
+              <Tooltip content="This will be visible to other users.">
+                <Icon active color="fg" name="info" padding={0.75} size="xs" tabIndex={0} />
+              </Tooltip>
+            </HStack>
           }
         />
       </VStack>
@@ -354,7 +352,7 @@ export const CustomLabel = () => {
         id="dob-tooltip"
         labelNode={
           <InputLabel htmlFor="dob-tooltip">
-            <HStack alignItems="center" gap={1}>
+            <HStack alignItems="center">
               Date of birth
               <Tooltip content="This will be visible to other users.">
                 <Icon active color="fg" name="info" size="xs" tabIndex={0} />